Create SPR Workflow package template

Create SPR Workflow package template

sprthis(wfName = "SPRtest", analysis, path = tempdir(), baseurl = "<baseurl>")

Arguments

wfName
character vector with the new SPR Workflow package name.

analysis
character vector withe the analysis name.

path
A path. If it exists, it is used. If it does not exist, it will be used the temporary path..

baseurl
baseurl for the package

Value

Path to the newly created package.

Examples

sprthis(wfName = "SPRtest", analysis = "SPRtest", path = tempdir())

#> ✔ Creating '/tmp/Rtmpe3Xu3w/SPRtest/'

#> ✔ Setting active project to '/tmp/Rtmpe3Xu3w/SPRtest'

#> ✔ Creating 'R/'

#> ✔ Writing 'DESCRIPTION'

#> Package: SPRtest
#> Title: SPRtest
#> Version: 0.9.0
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: This package provides a pre-configured
#>     workflow and reporting template for SPRtest.
#> License: Artistic-2.0
#> URL: https://github.com/systemPipeR/SPRtest
#> Imports:
#>     systemPipeR (>= 1.25.0)
#> Suggests:
#>     BiocStyle,
#>     knitr,
#>     rmarkdown
#> VignetteBuilder:
#>     knitr
#> biocViews: Infrastructure, ...
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.1.1
#> SystemRequirements: SPRtest can be used to run external
#>     command-line software, but the corresponding tool needs to be
#>     installed on a system.

#> ✔ Writing 'NAMESPACE'

#> ✔ Setting active project to '<no active project>'

#> ✔ Setting active project to '/tmp/Rtmpe3Xu3w/SPRtest'

#> ✔ Adding 'testthat' to Suggests field in DESCRIPTION

#> ✔ Setting Config/testthat/edition field in DESCRIPTION to '3'

#> ✔ Creating 'tests/testthat/'

#> ✔ Writing 'tests/testthat.R'

#> ✔ Writing 'tests/testthat/test-SPRtest.R'

#> ● Edit 'tests/testthat/test-SPRtest.R'

#> [1] "/tmp/Rtmpe3Xu3w/SPRtest"